[CS-1855] - Create webProvider singleton
